fix some bug
git-svn-id: http://bbs.cs.nthu.edu.tw/svn/maplebbs/trunk@301 bd1103fd-f9b1-db11-a66c-000d613fc1d9

+++ util/account.c
@@ -712,7 +712,7 @@ class_parse(key)
       return CH_END;
     }
 
-    chx[chn++] = chp = (ClassHeader *) malloc(sizeof(ClassHeader) + count * sizeof(short));
+    chx[chn++] = chp = (ClassHeader *) malloc(sizeof(ClassHeader) + count * sizeof(int));
     memset(chp->title, 0, CH_TTLEN);
     strcpy(chp->title, key);
 
@@ -799,7 +799,7 @@ class_sort(cmp)
   bhead = bp = bshm->bcache;
   btail = bp + max;
 
-  chp = (ClassHeader *) malloc(sizeof(ClassHeader) + max * sizeof(short));
+  chp = (ClassHeader *) malloc(sizeof(ClassHeader) + max * sizeof(int));
 
   for (i = j = 0; i < max; i++, bp++)
   {
@@ -809,7 +809,7 @@ class_sort(cmp)
 
   chp->count = j;
 
-  qsort(chp->chno, j, sizeof(short), cmp);
+  qsort(chp->chno, j, sizeof(int), cmp);
 
   memset(chp->title, 0, CH_TTLEN);
   strcpy(chp->title, "Boards");
@@ -822,7 +822,7 @@ class_image()
 {
   int i, times;
   FILE *fp;
-  short len, pos[CH_MAX];
+  int len, pos[CH_MAX];
   ClassHeader *chp;
 
   for (times = 2; times > 0; times--)	/* itoc.010413: 產生二份 class image */
@@ -834,23 +834,23 @@ class_image()
     if (chn < 2)		/* lkchu.990106: 尚沒有分類 */
       return;
 
-    len = sizeof(short) * (chn + 1);
+    len = sizeof(int) * (chn + 1);
 
     for (i = 0; i < chn; i++)
     {
       pos[i] = len;
-      len += CH_TTLEN + chx[i]->count * sizeof(short);
+      len += CH_TTLEN + chx[i]->count * sizeof(int);
     }
 
     pos[i++] = len;
 
     if (fp = fopen(CLASS_RUNFILE, "w"))
     {
-      fwrite(pos, sizeof(short), i, fp);
+      fwrite(pos, sizeof(int), i, fp);
       for (i = 0; i < chn; i++)
       {
 	chp = chx[i];
-	fwrite(chp->title, 1, CH_TTLEN + chp->count * sizeof(short), fp);
+	fwrite(chp->title, 1, CH_TTLEN + chp->count * sizeof(int), fp);
 	free(chp);
       }
       fclose(fp);
